fixed casting error in camera_functions

+++ camera_functions.cpp
@@ -17,7 +17,7 @@ float set_error(int *linePoints, int *whitePixels) {
 
     // change the error variable to the average error
     if (whitePixels > 0) {
-        error = error / whitePixels;
+        error = error / (float)whitePixels;
     }
 
     return error;
